Move Music Manager to separate module
